build: get OpenBSD working again with a go1.4 tree
I ran an OpenBSD 1.4.1 build by hand with gomote, ran gomote gettar to
get the built tree, and put it on GCS.

This CL then adds the URL of that Go 1.4 pre-built tree to the OpenBSD
builder's configuration, and modifies the coordinator to tell the
buildlet to pull extract the tarball from that URL to the "go1.4"
directory before the build starts.

Aside: I'm taking advantage of the build repo's new freedom in
depending on outside libraries and using Camlistore's syncutil.Group.

Change-Id: I2d6e416d2220d2d34c630fed50344bf2de4ae710
Reviewed-on: https://go-review.googlesource.com/3114
Reviewed-by: Andrew Gerrand <adg@golang.org>