cmd/coordinator/buildongce: use CoreOS stable
Alpha introduced a problem with "docker logs":

https://groups.google.com/forum/#!topic/coreos-user/VhqeIjkO2VA

So, revert for now.  No need for bleeding edge.

Also, be sure to make the coordinator's service declaration restarts on
failure, even if there are many errors in a row.  (I learned this the
hard way with another project and just realized this service file
omits the StartLimitInterval=0 line to disable the rate limit)

Change-Id: I9717e3b665bd4540d3cf4921018b619af1539178
Reviewed-on: https://go-review.googlesource.com/14771
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>