cmd/coordinator: avoid ambiguous branch names
When trying to resolve a branch name, if there is a remote or tag
of the same name, git cannot resolve it and issues an warning that
"refname 'BRANCH' is ambiguous."  This was discovered in gccgo's
watcher process on the branch 'asan' which is also the name of a tag
in the gcc repo.  To disambiguate between a remote and a tag, any
refname can be prepended with "heads/" or "tags/", respectively.

Change-Id: I16679f5249a5f22df59ab3a0ed49459bbbf5367b
Reviewed-on: https://go-review.googlesource.com/16083
Reviewed-by: Andrew Gerrand <adg@golang.org>