cmd/coordinator: use unambiguous change triple
There are several places where the coordinator passes a CL's Change-ID
back to the Gerrit API. However, these APIs will fail if there are
multiple changes with the same Change-ID (for example, if there's a
cherry-pick). One consequence of this is that the coordinator can't
post comments on changes with ambiguous Change-IDs, which means it
can't post trybot status or update the TryBot-Result label.

Because of this ambiguity, the Gerrit APIs also accept a more complete
and unique triple of (project, branch, Change-ID). This commit adds a
ChangeTriple method to tryKey that returns this triple and modifies
all API calls to use this, as well as a few places on the status
pages.

Fixes golang/go#13331.

Change-Id: Ia13f66a5cb27a14dfc4c95b7e460110d2dbefe62
Reviewed-on: https://go-review.googlesource.com/17071
Reviewed-by: Andrew Gerrand <adg@golang.org>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>