cmd/coordinator: log event spans with associated success, not stop/start text
Now events and phases of the build are represented by a span, with a
stop and start, and each has timing and a success status.

These will visualize nicer (later) and will be able to be logged and
queried in Cloud Datastore + BigQuery.

Also, remove the overloaded and not-to-be-used "rev" parameter from
the GetBuildlet interface. An upcoming change disconnects the
buildlet's lifetime from that of a single build, so the rev isn't even
known at the time of the GetBuildlet call; the buildlet will be handed
out to the most worthy current waiter (and the rev might not have even
existed at the tim the GetBuildlet call started)

Updates golang/go#12669

Change-Id: I02b28fea8bbbc16dc95c1ea0507418be01aa8ca8
Reviewed-on: https://go-review.googlesource.com/21570
Reviewed-by: Andrew Gerrand <adg@golang.org>