cmd/coordinator: be less conservative logging branches
Rather than being clever by logging from the branch head to the merge
base, it should be fine to 'git log' everything from the branch head to
the beginning of time. This only happens when the watcher starts; after
that it does a log from the branch head to the last known head.

Fixes golang/go#16446

Change-Id: I055be9f255423444ee9b78030969290af6b4fe9f
Reviewed-on: https://go-review.googlesource.com/25086
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>