cmd/coordinator: don't treat 'dist test list' remote errors as temp errors
If the buildlet interface reports that "dist test list" was a
"remoteErr" (the remote side successfully ran the command and relayed
its exit status to us, and it was a failure), don't treat that as an
"exec error" (that is, a communication or temporary error).

This is what tied up all the builders forever the other night when the
arm and ppc64 builds were broken, unable to run the go binary, since
they were forever retrying and not making progress. They should've
reported an error to the dashboard.

This was the fix deployed last night.

Change-Id: I08ce15d87eac5374eb907d4a7c72278106ecaba9
Reviewed-on: https://go-review.googlesource.com/27655
Reviewed-by: Andrew Gerrand <adg@golang.org>