cmd/coordinator: fix bug in gomote reverse proxying
ReverseProxy.ServeHTTP expects a server *http.Request.

The gomote proxy in the coordinator was making a client *http.Request
(which has different semantics for ContentLength depending on whether
it's client or server) and that was confusing the ReverseProxy code.

Set the next (server) request's ContentLength explicitly to the
original incoming Request.ContentLength. It was zero before, which
used to be safely okay, but not safe as of
https://go-review.googlesource.com/28412

Change-Id: I96d5c2b992d223372f9a01d268d3a419ccdc0d95
Reviewed-on: https://go-review.googlesource.com/28751
Reviewed-by: Chris Broadfoot <cbro@golang.org>