devapp: serve static files outside of App Engine
Currently we rely on app.yaml for routing/serving static files in the
project. But this file isn't available if we are not running on App
Engine, so we need a different solution. Add a very basic directory
server. I'm not happy that we need N routes to serve N static files,
or that we can't reuse existing pieces but if I understand correctly
to use http.FileServer we'd need to have it handle the root route to
route all requests in a directory to the right handler.

Make the dashboard visible outside of App Engine by providing dummy
login/logout URL's, even though the functionality is unimplemented -
this makes the dashboard accessible and changes the error from a 500
(blocking the dashboard) to a 404 when you try to log in.

Change-Id: I2a52ff73dc68705962ecc64810ad352988097c06
Reviewed-on: https://go-review.googlesource.com/36319
Reviewed-by: Quentin Smith <quentin@golang.org>