cmd/gitmirror, gerrit: move GetProjects to gerrit
This adds more utility to the Gerrit client and lets us reuse the
Gerrit API logic, instead of inventing it on an adhoc basis inside
the gitmirror code.

I added a test that this deserializes the correct data from the
Gerrit API for go.googlesource.com (and then removed it).

Change-Id: I449594b0257ed15d9b5aa7aa1eda7fb371fad6c1
Reviewed-on: https://go-review.googlesource.com/37252
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>