all: add linux-arm64-packet builders
The server has 96 cores and 128 GB of RAM, so run 20 builders on it.

This adds the Dockerfile each builder runs in, and the machinery to
run 20 copies of the Docker container.

Also reduce some logspam in the buildlet's shutdown.

(The dashboard/builders.go entry for this builder was added previously
 in https://golang.org/cl/39851)

Fixes golang/go#19929

Change-Id: I8537eee52c002dc9efcadcfb7e78b3a5db07ae44
Reviewed-on: https://go-review.googlesource.com/40392
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>