all: cleanup. update README, delete old stuff
The watcher is dead. It was renamed to gitmirror, but its
"watcher-world" environment is no longer relevant, since it's deployed
as a Kubernetes container now.

And the linux-x86-std VM configuration is old. We only use
linux-x86-std-kube on Kubernetes now.

Remove some gccgo stuff too, since that bitrot ages ago.

Change-Id: I74038961dee818b70a5e33c949a64a210bb1abb7
Reviewed-on: https://go-review.googlesource.com/40398
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>