env/darwin, cmd/{buildlet,buildlet/stage0,makemac}: clean MacStadium start-up
The Mac builders were a bit weird in how they started up and refreshed
their buildlet binaries. (Problematically, they DIDN'T refresh their
buildlet binaries properly compared to all the other environments)

This makes them mostly like the other environments and documents the
process more. Instead of booting and downloading a weird ltitle shell
script that only downloads the buildlet binary once, it now downloads
the stage0 binary instead, and the stage0 binary handles the looping
and the smart conditional re-downloading of the buildlet binary per
build.

This also adds a --destroy-all flag to makemac (used to test the new
setup) and fixes a bug found in makemac while testing: it tried to
start twenty 10.8 instances because it didn't account for their boot
time when considering what was needed next. Now instead of just using
what's currently connected to the coordinator, it uses the versions
of the VMs created, even if they're booting and not connected yet.

Change-Id: Ie76b35d3a8c032136568bc0812d8c7dd54e7fa28
Reviewed-on: https://go-review.googlesource.com/53416
Reviewed-by: Herbie Ong <herbie@google.com>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>