cmd/gerritbot: don’t create more than one Gerrit change for a PR
Due to maintner’s async nature, there is a period of time where
a Gerrit change may exist for a Pull Request but it is not reflected
in the corpus. Account for this by keeping an in-memory map of PRs
with stale data, and query Gerrit (should that map be incorrect due
to a restart) before creating a new change.

Update: golang/go#18517
Change-Id: Ife142697b15269caca088c3616c4acd4b227bfeb
Reviewed-on: https://go-review.googlesource.com/80840
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>