cmd/coordinator, dashboard: add mac race builder, skip redundant trybot tests
This adds a darwin-amd64-race builder. Of the four platforms that
support the race detector, this was the only one missing a builder.

To make up for the extra work, skip some redundant tests that the Mac
(and other) builds do. The test directory is really slow, and doesn't
have platform-specific code. We have amd64 & 386 test coverage for
that directory via the Linux builders, which is fine. No need to do it
on some of the heavier VM builders, or for trybots.

Fixes golang/go#17674

Change-Id: I888119b3fdd28884b87a59d78746ab1fef6960c7
Reviewed-on: https://go-review.googlesource.com/82395
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>