cmd/gerritbot: use merge --squash instead of resetting commits
Previously, it would reset commits based on the number reported
by GitHub. This would cause issues if there was a merge commit
due to them having more than one parent. So GitHub would report
three commits, but one of them was a merge commit so the number
would be off and it would push the wrong change.

Change-Id: I1d16c7cabdb96c9cef45ce0b802c1c5abd12e2de
Reviewed-on: https://go-review.googlesource.com/94636
Reviewed-by: Jonathan Nieder <jrn@google.com>
Reviewed-by: Andrew Bonventre <andybons@golang.org>