cmd/coordinator: style the "trybot status" page
Load style.css so we share some styles with the Go Build Coordinator
homepage. Add some custom styles so text in tables is a little bigger
- there's less of it on the trybot page.

Instead of formatting to the nearest tenth of a minute, we can round
to the nearest second - Duration.Round() is now two Go versions old
- and use more natural duration formatting.

Instead of issuing many small writes to the ResponseWriter, write HTML
to a buffer and then write it to the ResponseWriter all at once. (We
should just use a template but that's a bigger project.)

Add a "/try-dev" handler you can enable with the `-dev` build tag
which makes it easy to view and test the "trybot status" page locally,
without requiring a builder.

Change-Id: I28617d02ed857c28d2bb2d9ccfb05ca9dc572212
Reviewed-on: https://go-review.googlesource.com/103870
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>