maintner: change type/name of GerritCL meta fields, ensure Meta always non-nil
They were of type GitCommit, but we want to hang the useful methods
off the existing GerritMeta wrapper.

This was the plan earlier (when the GerritMeta type was added), but
that earlier CL didn't go all the way and convert callers.

This will likely break some maintner-using code, but there aren't many
callers.

Also, document & enforce that Meta & Commit are always non-nil when
enumerating CLs and when looking up CLs. Previously, two deleted CLs
didn't have this property, as the Gerrit server was returning a ref
for a version of the commit, but no meta ref. Filter those out.

Fixes golang/go#22060
Fixes golang/go#21984

Change-Id: I67727977c7a25e750f51a6ff7f3e00aaf960ef9f
Reviewed-on: https://go-review.googlesource.com/107296
Reviewed-by: Dmitri Shuralyov <dmitri@shuralyov.com>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>