env/linux-x86-nacl: update builders to unreleased pepper_67
The nacl image hadn't been updated in 2+ years and it needed to be
updated as part of rolling out the new COS-based builders.

But no released version works for us yet; we were getting the same
errors as in golang/go#23836 ("Signal 11 from untrusted code")

We were getting lucky that it was working with an ancient (pepper_34?)
version, but I was unable to get those working again either.

Rolling forward is better anyway, as we haven't had a Dockerfile
reflecting reality for this builder for 2+ years.

This is the same version used in playground in CL 101735, which said:

> playground: update NaCl to trunk.544461
>
> This pulls in https://crrev.com/c/962675, which fixes the
> underlying issue of NaCl mishandling signals during a SIGSEGV.

Updates golang/go#23836
Updates golang/go#25108

Change-Id: I187042af71a1249e84ce2070aa8039a88d2c02c2
Reviewed-on: https://go-review.googlesource.com/112735
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Andrew Bonventre <andybons@golang.org>