cmd/racebuild: improve reproducibility of .syso builds
Use a known version of MinGW on Windows. (On other platforms, we use
whatever compiler is already installed on the system image.)

Test at a well-defined Go commit.

Record the Go and LLVM commits for each platform independently.
Extracting the C++ toolchain version is left for future work, but it
should be recoverable from the resulting .syso file anyway.

Change-Id: I9af1d2a6f540a4d276b87074864564bf989e4731
Reviewed-on: https://go-review.googlesource.com/115375
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>