cmd/gopherbot: add has-cl label to issues with CL
For contributors looking for new issues to contribute to it can be
difficult to find issues that need a fix and don't already have a fix
being considered. There are several labels that help guide the way
already, like NeedsFix, HelpWanted. But many issues with this label will
already have a CL. There's currently no way to identify with a GitHub
search which issues have CLs and which do not. It'd be helpful to be
able to search for issues that do not have a CL and a label on issues
that have a CL is a simple method to distinguishing issues that do or do
not have a CL.

Fixes #26494