cmd/gerritbot: update gerritChangeRE for new Gerrit output format
Gerrit has changed the format it uses for the Change URL in its output.
It used to be something like:

	remote: SUCCESS
	remote:
	remote: New Changes:
	remote:   https://go-review.googlesource.com/#/c/dl/+/134117 remove blank line from codereview.cfg

Now it is:

	remote: SUCCESS
	remote:
	remote: New Changes:
	remote:   https://go-review.googlesource.com/c/dl/+/134117 remove blank line from codereview.cfg

(The '#' path element in the Change URL is removed.)

Update the regexp to match, and add a test for it.

This fix will likely cause the previously unposted "This PR has been
imported to Gerrit for code review" comments to get posted during
next pull request activity, since they haven't already been.

Fixes golang/go#27561.
Updates golang/go#27504.

Change-Id: I37a75ddb8d879017ebd887c651ca74a6ad6b892d
Reviewed-on: https://go-review.googlesource.com/134175
Reviewed-by: Andrew Bonventre <andybons@golang.org>