dashboard, env/openbsd: add OpenBSD 6.4 snapshot builders
This change adds amd64 and 386 OpenBSD 6.4 snapshot builders.
The snapshots are from October 4, 2018:

	/pub/OpenBSD/snapshots/amd64/BUILDINFO:
	Build date: 1538669162 - Thu Oct  4 16:06:02 UTC 2018

	/pub/OpenBSD/snapshots/i386/BUILDINFO:
	Build date: 1538681766 - Thu Oct  4 19:36:06 UTC 2018

These are meant to be replaced with the final version of OpenBSD 6.4
when it is released.

Modify make.bash to support fetching snapshot versions.

The amd64 architecture image was tested with cmd/debugnewvm
and passed all tests successfully.

Updates golang/go#26833.

Change-Id: Iea456605ace08f25d4eba37f9a962e6b9af95369
Reviewed-on: https://go-review.googlesource.com/c/139857
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>