cmd/coordinator: run subrepo trybots against recent Go release branches
This is the coordinator half of the change to make the coordinator
test subrepos against two prior releases of Go.

The next change that's required is for the maintnerd API server to
return the name of the past two releases.

For now this change is a no-op. But once maintnerd starts returning
more data, then this change will start causing causing two more
linux-amd64 builders per subrepo trybot.

Updates golang/go#17626

Change-Id: I1cedbc2e4eee51fb003c8bcc8e072fd10717a833
Reviewed-on: https://go-review.googlesource.com/c/143538
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>