cmd/release: fix tour after rename, do less tour work for Go 1.12+ releases
The tour binary was renamed from gotour to tour in CL 141857 as part
of golang/go#28163, but cmd/release wasn't updated to know about the new
name. This fixes that.

Additionally, don't push tour source to the buildlet if the releaselet
isn't going to include it anyway.

Tested that running it with a Go 1.11 release includes the tour:

   $ release -rev release-branch.go1.11 -tools release-branch.go1.11 \
             -target=linux-amd64 -skip_tests -version=go1.11.99 -watch

And with Go 1.12 does not:

   $ release -rev master -tools master \
             -target=linux-amd64 -skip_tests -version=go1.12.99 -watch

Updates golang/go#28163
Updates golang/go#27151

Change-Id: I246fc274c9dea81a24d27322bcbbe5a59e682a5e
Reviewed-on: https://go-review.googlesource.com/c/143537
Reviewed-by: Katie Hockman <katie@golang.org>